fixed linux stuff

+++ src/Simulator.cpp
@@ -1921,7 +1921,7 @@ std::vector<bool> Simulator::createMarkovChainSolver(const std::vector<DiceIdx>&
 
     // get optimized rerollers:
     float temp_erg = 0.0F;
-    size_t temp_rerollers = std::min(roll_atempts, 2ULL);
+    size_t temp_rerollers = std::min(roll_atempts, static_cast<size_t>(2));
     if (rerolls >= 2)
     {
         temp_rerollers = 2;
@@ -2357,7 +2357,7 @@ std::vector<bool> Simulator::createMarkovChainSolverEigen(const std::vector<Dice
 
     // get optimized rerollers:
     float temp_erg = 0.0F;
-    size_t temp_rerollers = std::min(roll_atempts, 2ULL);
+    size_t temp_rerollers = std::min(roll_atempts, static_cast<size_t>(2));
     if (rerolls >= 2)
     {
         temp_rerollers = 2;
